minor refactoring to reduce duplication in optimised Trie nodes
