+++ brjs-core/src/main/java/org/bladerunnerjs/utility/trie/node/OptimisedTrieRootNode.java
@@ -1,7 +1,7 @@
 package org.bladerunnerjs.utility.trie.node;
 
 
-public class OptimisedTrieRootNode<T> implements TrieNode<T>
+public class OptimisedTrieRootNode<T> extends AbstractOptimisedTrieNode<T>
 {
 
 	TrieNode<T>[] children;
@@ -14,61 +14,12 @@ public class OptimisedTrieRootNode<T> implements TrieNode<T>
 	@Override
 	public TrieNode<T> getNextNode(char character)
 	{
-		for (TrieNode<T> trieNode : children) {
-			if (trieNode != null && trieNode.getChar() == character) {
-				return trieNode;
-			}
-		}
-		return null;
-	}
-	
-	@Override
-	public TrieNode<T> getOrCreateNextNode(char character)
-	{
-		return getNextNode(character);
+		return getNextNode(children, character);
 	}
 	
 	@Override
 	public char getChar()
 	{
 		return '\u0000';
-	}
-	
-	@Override
-	public T getValue()
-	{
-		unsupportedOptimisedTrieMethod();
-    	return null;
-	}
-
-	@Override
-    public void setValue(T value)
-    {
-    	unsupportedOptimisedTrieMethod();
-    }
-    
-    @Override
- 	public TrieNode<T>[] getChildren()
- 	{
-    	unsupportedOptimisedTrieMethod();
-    	return null;
- 	}
-
-    @Override
- 	public int compareTo(TrieNode<T> compareNode)
- 	{
-    	unsupportedOptimisedTrieMethod();
-    	return 0;
- 	}
-
-	@Override
-	public int size()
-	{
-		unsupportedOptimisedTrieMethod();
-		return -1;
-	}
-	
-	private void unsupportedOptimisedTrieMethod() {
-		throw new RuntimeException("This is an optimised TrieNode and doesn't support this method.");
 	}
 }
