+++ brjs-core/src/main/java/org/bladerunnerjs/utility/trie/node/OptimisedTrieTrunkNode.java
@@ -1,7 +1,7 @@
 package org.bladerunnerjs.utility.trie.node;
 
 
-public class OptimisedTrieTrunkNode<T> implements TrieNode<T>
+public class OptimisedTrieTrunkNode<T> extends AbstractOptimisedTrieNode<T>
 {
 
 	private char character;
@@ -16,18 +16,7 @@ public class OptimisedTrieTrunkNode<T> implements TrieNode<T>
 	@Override
 	public TrieNode<T> getNextNode(char character)
 	{
-		for (TrieNode<T> trieNode : children) {
-			if (trieNode != null && trieNode.getChar() == character) {
-				return trieNode;
-			}
-		}
-		return null;
-	}
-	
-	@Override
-	public TrieNode<T> getOrCreateNextNode(char character)
-	{
-		return getNextNode(character);
+		return getNextNode(children, character);
 	}
 	
 	@Override
@@ -36,41 +25,4 @@ public class OptimisedTrieTrunkNode<T> implements TrieNode<T>
 		return character;
 	}
 	
-    @Override
-    public T getValue()
-    {
-    	unsupportedOptimisedTrieMethod();
-    	return null;
-    }
-    
-    @Override
-    public void setValue(T value)
-    {
-    	unsupportedOptimisedTrieMethod();
-    }
-    
-    @Override
- 	public TrieNode<T>[] getChildren()
- 	{
-    	unsupportedOptimisedTrieMethod();
-    	return null;
- 	}
-
-    @Override
- 	public int compareTo(TrieNode<T> compareNode)
- 	{
-    	unsupportedOptimisedTrieMethod();
-    	return 0;
- 	}
-
-	@Override
-	public int size()
-	{
-		unsupportedOptimisedTrieMethod();
-		return -1;
-	}
-	
-	private void unsupportedOptimisedTrieMethod() {
-		throw new RuntimeException("This is an optimised TrieNode and doesn't support this method.");
-	}
 }
